Villager queries
I'm finding I'm bumping up against the minutae of modeling the data from json
I still want to hold off on having a db for as long
as possilbe but right now it's really fiddly to do
these deep joins. Especially with villagers and the inculsion
of week/season rosters. Sheesh